Python 3.10 support, + FreeBSD 13 support
Mostly focuses on some revamping of the dependencies. We now use more
complicated dependency requirements, based on the Python version /
environment.

+++ scripts/teamcity/20.test.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 
+set -u -x
 PYTHON="${PYTHON_CMD:-python3}"
 
 # remove previous environment
@@ -13,7 +14,11 @@ if [ -d .env ]; then
         ;;
     esac
     ${PYTHON} -m pip freeze > to_remove.txt
-    ${PYTHON} -m pip uninstall -r to_remove.txt -y
+
+    if [ -s to_remove.txt ]; then
+        ${PYTHON} -m pip uninstall -r to_remove.txt -y
+    fi
+
     deactivate
     rm -Rf .env
 fi
@@ -30,8 +35,16 @@ esac
 
 # first remove system then user
 ${PYTHON} -m pip uninstall -r dev-requirements.txt -y
-${PYTHON} -m pip uninstall -r dev-requirements.txt -y
 ${PYTHON} -m pip install --upgrade pip
+${PYTHON} -m pip install --upgrade wheel
+${PYTHON} -m pip install --upgrade "setuptools<=58.4"
 ${PYTHON} -m pip install -r dev-requirements.txt
-${PYTHON} -m pip install --upgrade setuptools wheel
-${PYTHON} setup.py test  --addopts "--junitxml=${JUNIT_XML_FILE}"
+
+TEST_OPTS="-s"
+if [[ ! -z ${JUNIT_XML_FILE-} ]]
+then
+    TEST_OPTS+=" --junitxml=${JUNIT_XML_FILE}"
+fi
+
+echo "Invoking pytest with --addopts '${TEST_OPTS}'"
+${PYTHON} setup.py test  --addopts "${TEST_OPTS}"
